patch-image tool needs doc cleanup and bugfixes
Issues: Addresses Issue #4

Problem: The patch image tool needs better functional structure and
better documentation for usage. The startup script and it's functions
need merging with those in perforce

Analysis: Added info to the main README.rst in this repo to demonstrate
usage of the patch image tool. Merged in the new startup script changes.

Tests: