set the expected iface name for >= 13.x
Issues:
Fixes #54

Problem: The logical name of the NIC changed as a function
of the version of the BIGIP.  If the BIGIP is >= 13.x, then
the interface is named "mgmt".

Analysis:  This affected the automated onboarding script. To
handle the case I've added a few functions.  See function comments
for details.

Tests: To test I ran the "startup" script on a system with the new
functions.  The hostname was correctly assigned after the new functions
were defined and exported.   To properly test negatively it is necessary
to run the new version of the code against an "unpatched image" with
versions less than 13.  This test will be more efficiently run once this
commit is accepted so it has NOT been run yet.