Fixes #62
This patch provides a script and associated JSON blob key that allows an administrator to
provide a wc_notify url to the startup script in much the same way that they would if they
were using cloud-init

os-functions/openstack-wait-handler.sh

In v13, using clout-init, this functionality is possible to add and allows an admin to
use Heat to accurately notify the admin when the bigip VM is complete. In versions <13
though, this is not possible. This patch adds functionality though so that it is possible
and therefore makes heat successfully return "complete" only when the VM is actually completely
configured and not earlier.