setup.py install does not require openstack clients
Issues:
Fixes #60

Problem:
The setup.py installation of this repo does not require the proper
python client libraries for glance and keystone. This is currently done
in the f5-openstack-heat template with an apt-get install (which is not
correct). We'll move it to the setup.py here and then update the heat
template to remove that installation. The other big point is that we are
installation from latest with apt-get, so we're unsure of what we're
getting when we do the heat template install. With the setup.py change,
we'll fix it to a particular version that is tested and works.

Analysis:
Added install_requires to have the keystone and glance client for
openstack. These are pinned to specific versions. Those versions will
change between libert and mitaka etc.

Tests:
Tested with heat deployment from f5-openstack-heat repo.